look at this example:
🦷 *Appointment Tomorrow!* 🦷

Hi {{UserName}}, ((here there should alwasy be a line break after a hello and name callout. does this make sense?
This is a reminder for your appointment with *Dr. Smile* tomorrow at *10:00 AM* for a dental check-up.

📍 {{ClinicAddress}}

To confirm, reply with "YES". To reschedule, reply with "NO" or call us at {{ClinicPhone}}.